Fix Windows cursor position query code
